requirements.txt
streamlit==1.39.0
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
numpy==2.1.1

google-genai>=0.10.0
tenacity==8.5.0

Pillow==10.4.0
moviepy==1.0.3
imageio==2.35.1
imageio-ffmpeg==0.5.1
